Made XP a double.

+++ app/src/main/java/io/prolabs/pro/algo/UserXp.java
@@ -4,13 +4,13 @@ package io.prolabs.pro.algo;
  * Created by Edmund on 2015-03-07.
  */
 public class UserXp {
-    private long totalXp;
+    private double totalXp;
 
-    public UserXp(long totalXp) {
+    public UserXp(double totalXp) {
         this.totalXp = totalXp;
     }
 
-    public long getTotalXp() {
+    public double getTotalXp() {
         return totalXp;
     }
 }
